feat: fixing autobahn tests for server, considering compression
